requirements.txt
coloredlogs~=15.0.1
aiogram~=2.15
tortoise-orm
tabulate~=0.8.9
matplotlib~=3.4.3
mplcyberpunk~=0.3.0
numpy~=1.21.4
pytz~=2021.3
aioschedule